fix build corner cases

+++ bin/build-go.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 # Downloads, builds, and installs a specified go version into your home
 # Installed as ~/go[version] and symlinked to ~/go
-# After install, just add to your path:
-# export PATH=$PATH:~/go/bin
+# After install, just prepend to your path:
+# export PATH=~/go/bin:$PATH
 
 set -e
 
@@ -11,7 +11,9 @@ if [ $1 ]; then
     if [ ! -f "go$1.src.tar.gz" ]; then
 	curl -LO "https://dl.google.com/go/go$1.src.tar.gz"
     fi
-    rm -rf go
+    if [ -d "go" ]; then
+        rm -rf go
+    fi
     tar -xvf "go$1.src.tar.gz" || true
 
     # build
@@ -31,6 +33,9 @@ if [ $1 ]; then
     rm -rf go/test
 
     # install
+    if [ -d "$GOROOT_FINAL" ]; then
+        rm -rf "$GOROOT_FINAL"
+    fi
     mv -v go $GOROOT_FINAL
     ln -snfv $GOROOT_FINAL ~/go
 else
